@@ -1,6 +1,6 @@
 {
   "name": "typed-rest-client",
-  "version": "1.8.11",
+  "version": "2.0.0",
   "description": "Node Rest and Http Clients for use with TypeScript",
   "main": "./RestClient.js",
   "scripts": {
@@ -27,20 +27,25 @@
   "bugs": {
     "url": "https://github.com/Microsoft/typed-rest-client/issues"
   },
+  "engines": {
+    "node": ">= 16.0.0"
+  },
   "homepage": "https://github.com/Microsoft/typed-rest-client#readme",
   "devDependencies": {
+    "@octokit/rest": "^20.1.1",
     "@types/mocha": "^2.2.44",
-    "@types/node": "^6.0.92",
+    "@types/node": "^20.11.0",
     "@types/shelljs": "0.7.4",
-    "mocha": "^6.2.3",
-    "nock": "9.6.1",
-    "semver": "4.3.3",
+    "mocha": "^10.4.0",
+    "nock": "13.5.4",
+    "semver": "^7.6.2",
     "shelljs": "^0.8.5",
-    "typescript": "3.1.5",
-    "@octokit/rest": "^16.43.2"
+    "typescript": "^5.4.5"
   },
   "dependencies": {
-    "qs": "^6.9.1",
+    "des.js": "^1.1.0",
+    "js-md4": "^0.3.2",
+    "qs": "^6.9.7",
     "tunnel": "0.0.6",
     "underscore": "^1.12.1"
   }
