@@ -7,8 +7,8 @@
     "typed-rest-client": "file:../../_build"
   },
   "devDependencies": {
-    "ts-node": "^10.9.1",
-    "typescript": "^2.4.2"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   },
   "scripts": {
     "start": "ts-node samples.ts"
