@@ -1,41 +1,32 @@
 {
-  "name": "typed-rest-client-react-sample",
-  "version": "1.0.0",
-  "description": "Typed Rest Client React Sample",
+  "name": "react-2",
   "private": true,
-  "dependencies": {
-    "assert": "^2.0.0",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "https-browserify": "^1.0.0",
-    "path-browserify": "^1.0.1",
-    "react": "16.5.0",
-    "react-dom": "16.5.0",
-    "react-scripts": "^5.0.1",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "typed-rest-client": "file:../../_build",
-    "url": "^0.11.0",
-    "util": "^0.12.4"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not op_mini all"
-  ],
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "typed-rest-client": "file:../../_build"
+  },
   "devDependencies": {
-    "@types/node": "^18.0.6",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "process": "^0.11.10",
-    "react-app-rewired": "^2.2.1",
-    "typescript": "^4.7.4",
-    "webpack": "^5.73.0"
+    "@originjs/vite-plugin-commonjs": "^1.0.3",
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.6",
+    "stream-browserify": "^3.0.0",
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0",
+    "vite-plugin-node-polyfills": "^0.22.0"
   }
 }
