@@ -20,7 +20,7 @@
     "html-webpack-plugin": "^5.5.0",
     "https-browserify": "^1.0.0",
     "path-browserify": "^1.0.1",
-    "webpack": "^5.73.0",
+    "webpack": "^5.91.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.9.3"
   },
@@ -31,7 +31,7 @@
     "process": "^0.11.10",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
-    "typed-rest-client": "^1.8.9",
+    "typed-rest-client": "file:../../_build",
     "url": "^0.11.0",
     "url-loader": "^4.1.1",
     "util": "^0.12.4"
