@@ -30,7 +30,7 @@ function oddpar(buf)
  */
 function expandkey(key56)
 {
-  var key64 = new Buffer(8);
+  var key64 = Buffer.alloc(8);
 
   key64[0] = key56[0] & 0xFE;
   key64[1] = ((key56[0] << 7) & 0xFF) | (key56[1] >> 1);
@@ -49,7 +49,7 @@ function expandkey(key56)
  */
 function bintohex(bin)
 {
-  var buf = (Buffer.isBuffer(buf) ? buf : new Buffer(bin, 'binary'));
+  var buf = (Buffer.isBuffer(buf) ? buf : Buffer.from(bin, 'binary'));
   var str = buf.toString('hex').toUpperCase();
   return zeroextend(str, 32);
 }
