@@ -4,6 +4,7 @@ var $ = require('./common');
 var lmhashbuf = require('./smbhash').lmhashbuf;
 var nthashbuf = require('./smbhash').nthashbuf;
 
+var desjs = require("des.js");
 
 function encodeType1(hostname, ntdomain) {
   hostname = hostname.toUpperCase();
@@ -12,7 +13,7 @@ function encodeType1(hostname, ntdomain) {
   var ntdomainlen = Buffer.byteLength(ntdomain, 'ascii');
 
   var pos = 0;
-  var buf = new Buffer(32 + hostnamelen + ntdomainlen);
+  var buf = Buffer.alloc(32 + hostnamelen + ntdomainlen);
 
   buf.write('NTLMSSP', pos, 7, 'ascii'); // byte protocol[8];
   pos += 7;
@@ -86,10 +87,10 @@ function encodeType3(username, hostname, ntdomain, nonce, password) {
   hostname = hostname.toUpperCase();
   ntdomain = ntdomain.toUpperCase();
 
-  var lmh = new Buffer(21);
+  var lmh = Buffer.alloc(21);
   lmhashbuf(password).copy(lmh);
   lmh.fill(0x00, 16); // null pad to 21 bytes
-  var nth = new Buffer(21);
+  var nth = Buffer.alloc(21);
   nthashbuf(password).copy(nth);
   nth.fill(0x00, 16); // null pad to 21 bytes
 
@@ -110,7 +111,7 @@ function encodeType3(username, hostname, ntdomain, nonce, password) {
 
   var pos = 0;
   var msg_len = 64 + ntdomainlen + usernamelen + hostnamelen + lmrlen + ntrlen;
-  var buf = new Buffer(msg_len);
+  var buf = Buffer.alloc(msg_len);
 
   buf.write('NTLMSSP', pos, 7, 'ascii'); // byte protocol[8];
   pos += 7;
@@ -189,12 +190,17 @@ function encodeType3(username, hostname, ntdomain, nonce, password) {
 
 function makeResponse(hash, nonce)
 {
-  var out = new Buffer(24);
+  var out = Buffer.alloc(24);
+
   for (var i = 0; i < 3; i++) {
     var keybuf = $.oddpar($.expandkey(hash.slice(i * 7, i * 7 + 7)));
-    var des = crypto.createCipheriv('DES-ECB', keybuf, '');
-    var str = des.update(nonce.toString('binary'), 'binary', 'binary');
-    out.write(str, i * 8, i * 8 + 8, 'binary');
+
+    var des = desjs.DES.create({type: 'encrypt', key: keybuf});
+    var magicKey = Buffer.from(nonce.toString('binary'));
+    var insertBuff = Buffer.from(des.update(magicKey));
+
+    out.fill(insertBuff, i * 8, i * 8 + 8, 'binary');
+    
   }
   return out;
 }
@@ -210,7 +216,7 @@ exports.challengeHeader = function (hostname, domain) {
 };
 
 exports.responseHeader = function (res, url, domain, username, password) {
-  var serverNonce = new Buffer((res.headers['www-authenticate'].match(/^NTLM\s+(.+?)(,|\s+|$)/) || [])[1], 'base64');
+  var serverNonce = Buffer.from((res.headers['www-authenticate'].match(/^NTLM\s+(.+?)(,|\s+|$)/) || [])[1], 'base64');
   var hostname = require('url').parse(url).hostname;
   return 'NTLM ' + exports.encodeType3(username, hostname, domain, exports.decodeType2(serverNonce), password).toString('base64')
 };
