@@ -1,5 +1,7 @@
-var crypto = require('crypto');
 var $ = require('./common');
+
+var jsmd4 = require("js-md4");
+var desjs = require("des.js");
 
 /*
  * Generate the LM Hash
@@ -11,7 +13,7 @@ function lmhashbuf(inputstr)
   var xl = Buffer.byteLength(x, 'ascii');
 
   /* null pad to 14 bytes */
-  var y = new Buffer(14);
+  var y = Buffer.alloc(14);
   y.write(x, 0, xl, 'ascii');
   y.fill(0, xl);
 
@@ -24,12 +26,13 @@ function lmhashbuf(inputstr)
   /* DES encrypt magic number "KGS!@#$%" to two
    * 8-byte ciphertexts, (ECB, no padding)
    */
-  var buf = new Buffer(16);
+  var buf = Buffer.alloc(16);
   var pos = 0;
   var cts = halves.forEach(function(z) {
-    var des = crypto.createCipheriv('DES-ECB', z, '');
-    var str = des.update('KGS!@#$%', 'binary', 'binary');
-    buf.write(str, pos, pos + 8, 'binary');
+    var des = desjs.DES.create({type: 'encrypt', key: z});
+    var magicKey = Buffer.from('KGS!@#$%', 'ascii');
+    var insertBuff = Buffer.from(des.update(magicKey));
+    buf.fill(insertBuff, pos, pos + 8, 'binary');
     pos += 8;
   });
 
@@ -41,10 +44,10 @@ function lmhashbuf(inputstr)
 function nthashbuf(str)
 {
   /* take MD4 hash of UCS-2 encoded password */
-  var ucs2 = new Buffer(str, 'ucs2');
-  var md4 = crypto.createHash('md4');
+  var ucs2 = Buffer.from(str, 'ucs2');
+  var md4 = jsmd4.create();
   md4.update(ucs2);
-  return new Buffer(md4.digest('binary'), 'binary');
+  return Buffer.from(md4.digest('binary'), 'binary');
 }
 
 function lmhash(is)
