@@ -86,7 +86,7 @@ function buildParamsStringifyOptions(queryParams: IRequestQueryParams): any  {
  * @param {string} charset? - optional; defaults to 'utf-8'
  * @return {Promise<string>}
  */
-export async function decompressGzippedContent(buffer: Buffer, charset?: string): Promise<string> {
+export async function decompressGzippedContent(buffer: Buffer, charset?: BufferEncoding): Promise<string> {
     return new Promise<string>(async (resolve, reject) => {
         zlib.gunzip(buffer, function (error, buffer) {
             if (error) {
@@ -128,16 +128,19 @@ export function buildProxyBypassRegexFromEnv(bypass : string) : RegExp {
  * @param {IHttpClientResponse} response
  * @return {string} - Content Encoding Charset; Default=utf-8
  */
-export function obtainContentCharset (response: IHttpClientResponse) : string {
+export function obtainContentCharset (response: IHttpClientResponse) : BufferEncoding {
   // Find the charset, if specified.
   // Search for the `charset=CHARSET` string, not including `;,\r\n`
   // Example: content-type: 'application/json;charset=utf-8'
   // |__ matches would be ['charset=utf-8', 'utf-8', index: 18, input: 'application/json; charset=utf-8']
   // |_____ matches[1] would have the charset :tada: , in our example it's utf-8
   // However, if the matches Array was empty or no charset found, 'utf-8' would be returned by default.
-  const nodeSupportedEncodings = ['ascii', 'utf8', 'utf16le', 'ucs2', 'base64', 'binary', 'hex'];
+  const nodeSupportedEncodings: BufferEncoding[] = ['ascii', 'utf8', 'utf16le', 'ucs2', 'base64', 'binary', 'hex'];
   const contentType: string = response.message.headers['content-type'] || '';
   const matches: (RegExpMatchArray|null) = contentType.match(/charset=([^;,\r\n]+)/i);
+  if (matches && matches[1] && nodeSupportedEncodings.indexOf(matches[1] as BufferEncoding) != -1) {
+    return matches[1] as BufferEncoding;
+  }
 
-  return (matches && matches[1] && nodeSupportedEncodings.indexOf(matches[1]) != -1) ? matches[1] : 'utf-8';
+  return 'utf-8';
 }
