@@ -1,15 +1,11 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import React from 'react';
 import { ClassExample } from './ClassExample';
 
-function App(): JSX.Element {
-    return (
-        <div className='app'>
-            <ClassExample />
-        </div>
-    );
+function App() {
+  return (
+    <div className='app'>
+      <ClassExample />
+    </div>
+  )
 }
 
-export default App;
+export default App
