@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import * as httpc from 'typed-rest-client/HttpClient';
+import httpc from 'typed-rest-client/HttpClient';
 
 export async function exampleGetWithHttp() {
     let result;
