@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import * as restc from 'typed-rest-client/RestClient';
+import restc from 'typed-rest-client/RestClient';
 
 export async function exampleGetWithRest() {
     let result;
@@ -9,7 +9,6 @@ export async function exampleGetWithRest() {
 
     try {
         const response = await client.get<any>('https://httpbin.org/get');
-
         result = 'Received response from: ' + response.result.url + ' using REST client'
     }
     catch (e: any) {
